pyproject.toml
[project]
name = "ollama-mcp-devops-server"
version = "2.0.0"
description = "Python-based MCP DevOps Tools Server with Ollama Integration"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "mupoese", email = "mupoese@users.noreply.github.com"},
]
keywords = ["mcp", "ollama", "devops", "ai", "llm", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
]
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "structlog>=23.2.0",
    "click>=8.1.7",
    "rich>=13.7.0",
]

[project.optional-dependencies]
devops = [
    "docker>=7.0.0",
    "kubernetes>=29.0.0",
    "python-terraform>=0.10.1",
    "GitPython>=3.1.40",
    "PyGithub>=2.1.1",
    "python-gitlab>=4.4.0",
]
cloud = [
    "boto3>=1.34.0",
    "google-cloud>=0.34.0",
    "azure-identity>=1.15.0",
    "azure-mgmt-core>=1.4.0",
]
monitoring = [
    "prometheus-client>=0.19.0",
    "elasticsearch>=8.12.0",
    "influxdb>=5.3.1",
]
security = [
    "cryptography>=42.0.0",
    "pyjwt>=2.8.0",
]
database = [
    "sqlalchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",
    "pymongo>=4.6.1",
    "redis>=5.0.1",
]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.2",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "black>=23.12.1",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
all = [
    "ollama-mcp-devops-server[devops,cloud,monitoring,security,database,dev]"
]

[project.scripts]
mcp-devops-server = "mcp_devops_server.main:cli"

[project.urls]
Homepage = "https://github.com/mupoese/Ollama-MCP-Server"
Repository = "https://github.com/mupoese/Ollama-MCP-Server.git"
Issues = "https://github.com/mupoese/Ollama-MCP-Server/issues"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_devops_server*"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["mcp_devops_server"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]